// src/card/card.types.d.ts
// prettier-ignore
export type Suite = "♠︎" | "♥︎" | "♣︎" | "♦︎" | "🃏";
// prettier-ignore
export type Value = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 13 | 15 | 17 | 21
// prettier-ignore
export type Rank = "A" | "2" | "3" | "4" | "5" | "6" | "7" | "8" | "9" | "10"| "J" | "Q" | "K" | "JOKER";